fix: dont automatically redirect from edit entry page
edit entry page had a check to try to go back if the entry to edit wasn't valid.
but there was no easy way to actually do that, and it ended up always going back,
even when the entry was valid.
this just removes that redirection, i.e. there is no more check and redirect -
it will simply count on the fact that in theory, it should never be possible
to be editing an invalid entry via navigating the application in the first place.